% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{centroid_dist}
\alias{centroid_dist}
\title{Compute distances between centroids}
\usage{
centroid_dist(sf)
}
\arguments{
\item{sf}{}
}
\value{
A distance matrix representing the pairwise distances between centroids.
}
\description{
Compute distances between centroids
}
