% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{efficiency_gap}
\alias{efficiency_gap}
\title{Calculate the Efficiency Gap of a districting plan
https://ballotpedia.org/Efficiency_gap}
\usage{
efficiency_gap(df)
}
\arguments{
\item{df}{}
}
\value{
The Efficency Gap of the districting plan.
}
\description{
Calculate the Efficiency Gap of a districting plan
https://ballotpedia.org/Efficiency_gap
}
