% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{polsby_popper}
\alias{polsby_popper}
\title{Calculate the Polsby-Popper compactness score}
\usage{
polsby_popper(sf)
}
\arguments{
\item{sf}{}
}
\value{
The Polsby-Popper compactness score of the districting plan.
}
\description{
https://en.wikipedia.org/wiki/Polsby-Popper_Test
}
